Add --browser-scan
