+++ pkg/config/browser.go
@@ -4,13 +4,15 @@ import (
 	"fmt"
 )
 
+// TODO - look at whether to collapse the json parsing down to the standard library. Want to be able to handle args being single string vs array etc...
+//      Also look at custom marhsalling: http://choly.ca/post/go-json-marshalling/
 type Browser struct {
-	Id       string
-	Name     string
-	Exe      string
-	Args     *string // TODO move to []string
-	IconPath *string
-	Hidden   bool
+	Id       string  `json:"id"`
+	Name     string  `json:"name"`
+	Exe      string  `json:"exe"`
+	Args     *string `json:"args"` // TODO move to []string to allow prompting for limited subset when matched
+	IconPath *string `json:"iconPath"`
+	Hidden   bool    `json:"hidden"`
 }
 
 // TODO - browser scan
